refactor: create camera module, move relevant items to it
