+++ src/main.rs
@@ -3,6 +3,7 @@
 
 mod animation;
 mod bundles;
+mod camera;
 mod control_display;
 mod event_scheduler;
 mod exorcism;
@@ -97,16 +98,16 @@ fn main() {
         .insert_resource(level_selection.clone())
         .insert_resource(level_transition::TransitionTo(level_selection))
         .insert_resource(resources::RewindSettings::NORMAL)
-        .insert_resource(resources::PlayZonePortion(0.75))
-        .add_startup_system(gameplay::transitions::spawn_camera)
+        .insert_resource(camera::PlayZonePortion(0.75))
+        .add_startup_system(camera::spawn_camera)
         .add_startup_system(gameplay::transitions::spawn_ui_root)
         .add_system_to_stage(CoreStage::PreUpdate, sugar::make_ui_visible)
         .add_enter_system(
             GameState::Gameplay,
-            gameplay::transitions::fit_camera_around_play_zone_padded,
+            camera::fit_camera_around_play_zone_padded,
         )
         .add_system(
-            gameplay::transitions::fit_camera_around_play_zone_padded
+            camera::fit_camera_around_play_zone_padded
                 .run_not_in_state(GameState::AssetLoading)
                 .run_on_event::<bevy::window::WindowResized>(),
         )
