+++ src/resources.rs
@@ -30,6 +30,3 @@ impl RewindSettings {
         hold_timer: None,
     };
 }
-
-#[derive(Copy, Clone, PartialEq, Debug, Default, Deref, DerefMut)]
-pub struct PlayZonePortion(pub f32);
